requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.13.4
frozenlist==1.4.0
fsspec==2024.3.1
gitdb==4.0.10
GitPython==3.1.36
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
gpt4all==2.6.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.286
langchain-community==0.0.34
langchain-core==0.1.46
langchain-google-genai==1.0.3
langchain-openai==0.1.4
langsmith==0.1.51
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.25.2
openai==1.23.6
orjson==3.10.1
packaging==23.2
pandas==2.1.0
Pillow==9.5.0
proto-plus==1.23.0
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.2
pypdf==3.16.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2024.4.16
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
SQLAlchemy==2.0.20
streamlit==1.26.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.4
validators==0.22.0
yarl==1.9.2
zipp==3.16.2
